+++ lib/src/_application/caching/registry.dart
@@ -1,24 +1,10 @@
-library plato.archives.models.caching.registry;
+library plato.archives.caching.registry;
 
 /// The [Registry] class...
-class Registry {
-  Map<String, dynamic> _registry;
-
-  static Registry _instance;
-
-  /// The [Registry] factory constructor...
-  factory Registry() => _instance ?? (_instance = new Registry._());
-
-  /// The [Registry] private constructor...
-  Registry._ () {
-    _registry = new Map<String, dynamic>();
-  }
-
+abstract class Registry<S, T> {
   /// The [register] method...
-  void register (String key, dynamic resource) {
-    _registry.putIfAbsent (key, () => resource);
-  }
+  void register (S key, T resource);
 
   /// The [unregister] method...
-  dynamic unregister (String key) => _registry.remove (key);
+  T unregister (S key);
 }