+++ lib/src/archive/non_retrievable_archive.dart
@@ -2,8 +2,8 @@ library plato.archives.error.retrieve_archives;
 
 import '../_application/error/plato_exception.dart';
 
-class ArchivesNotRetrievable extends PlatoException {
-  ArchivesNotRetrievable (
+class NonRetrievableArchive extends PlatoException {
+  NonRetrievableArchive (
     [String message = 'The list of archives were not retrievable.']
   ) : super (message);
 }