+++ lib/src/archive/retrieve_archives_service.dart
@@ -6,13 +6,21 @@ import 'dart:convert' show json;
 import 'package:angular/core.dart';
 import 'package:http/http.dart' show Client, Response;
 
-import 'archives_not_retrievable.dart';
+import '../enrollment/enrollment_factory.dart';
+import '../enrollment/faculty_enrollment.dart';
+
+import 'non_retrievable_archive.dart';
 
 const String _RETRIEVE_ARCHIVES_URI = '/plato/retrieve/archives';
 
 /// The [RetrieveArchivesService] class...
 @Injectable()
 class RetrieveArchivesService {
+  List<FacultyEnrollment> _facultyArchives;
+
+  List<FacultyEnrollment> get facultyArchives =>
+    new List<FacultyEnrollment>.from (_facultyArchives);
+
   final Client _http;
 
   static RetrieveArchivesService _instance;
@@ -22,18 +30,23 @@ class RetrieveArchivesService {
     _instance ?? (_instance = new RetrieveArchivesService._ (http));
 
   /// The [RetrieveArchivesService] private constructor...
-  RetrieveArchivesService._ (this._http);
+  RetrieveArchivesService._ (this._http) {
+    _facultyArchives = new List<FacultyEnrollment>();
+  }
 
-  /// The [retrieveArchives] method...
-  Future<void> retrieveArchives() async {
+  /// The [retrieveFacultyEnrollments] method...
+  Future<void> retrieveFacultyEnrollments() async {
     try {
       final Response retrieveArchivesResponse =
         await _http.get (_RETRIEVE_ARCHIVES_URI);
 
       List<Map<String, String>> rawArchives =
         (json.decode (retrieveArchivesResponse.body) as Map)['archives'];
+
+      _facultyArchives =
+        (new EnrollmentFactory()).createAll (rawArchives, 'faculty');
     } catch (_) {
-      throw new ArchivesNotRetrievable();
+      throw new NonRetrievableArchive();
     }
   }
 }