+++ lib/src/enrollment/faculty_enrollment.dart
@@ -1,14 +1,14 @@
 library plato.archives.models.enrollment.faculty;
 
-import 'enrollment.dart';
+import 'archive_enrollment.dart';
 
 /// The [FacultyEnrollment] class...
-class FacultyEnrollment extends Enrollment {
+class FacultyEnrollment extends ArchiveEnrollment {
   /// The [FacultyEnrollment] constructor...
   FacultyEnrollment (
     String id, String courseId, String courseName, String userId,
     bool available, {bool forArchive = true}
   ) : super (
-    id, courseId, courseName, userId, 'Instructor', available, forArchive: forArchive
+    id, courseId, courseName, userId, 'Instructor', available
   );
 }