+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # plato.archives:  A course archives viewer for Plato.
 
+## 0.16.1
+- Added code for enrollment factory.
+- Refactored some code for enrollment and archive modules.
+
 ## 0.16.0
 - Start of creating application caching layer.
 - Start of creating enrollments factory and related code.